window icon

+++ Nikesaurus/src/Systems/Render/sysRender.cpp
@@ -390,7 +390,7 @@ namespace NIKE {
 					throw std::exception();
 				}
 
-				const int texture_idx = std::distance(texture_binding_unit_map.begin(), texture_binding_unit_map.find(render_instances_texture[i].tex));
+				const int texture_idx = static_cast<int>(std::distance(texture_binding_unit_map.begin(), texture_binding_unit_map.find(render_instances_texture[i].tex)));
 
 				v.sampler_idx = static_cast<float>(texture_idx);
 				//v.sampler_idx = 0;
